# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoundTrip Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/emotions": {
      "post": {
        "tags": [
          "emotions"
        ],
        "summary": "Create Emotion",
        "operationId": "create_emotion_api_v1_emotions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmotionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmotionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "emotions"
        ],
        "summary": "List Emotions",
        "operationId": "list_emotions_api_v1_emotions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmotionRead"
                  },
                  "title": "Response List Emotions Api V1 Emotions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emotions/{emotion_id}": {
      "get": {
        "tags": [
          "emotions"
        ],
        "summary": "Get Emotion",
        "operationId": "get_emotion_api_v1_emotions__emotion_id__get",
        "parameters": [
          {
            "name": "emotion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Emotion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmotionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "emotions"
        ],
        "summary": "Update Emotion",
        "operationId": "update_emotion_api_v1_emotions__emotion_id__put",
        "parameters": [
          {
            "name": "emotion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Emotion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmotionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmotionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "emotions"
        ],
        "summary": "Delete Emotion",
        "operationId": "delete_emotion_api_v1_emotions__emotion_id__delete",
        "parameters": [
          {
            "name": "emotion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Emotion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/geographies": {
      "post": {
        "tags": [
          "geographies"
        ],
        "summary": "Create Geography",
        "operationId": "create_geography_api_v1_geographies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeographyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "geographies"
        ],
        "summary": "List Geographies",
        "operationId": "list_geographies_api_v1_geographies_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeographyRead"
                  },
                  "title": "Response List Geographies Api V1 Geographies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/geographies/{geography_id}": {
      "get": {
        "tags": [
          "geographies"
        ],
        "summary": "Get Geography",
        "operationId": "get_geography_api_v1_geographies__geography_id__get",
        "parameters": [
          {
            "name": "geography_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Geography Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "geographies"
        ],
        "summary": "Update Geography",
        "operationId": "update_geography_api_v1_geographies__geography_id__put",
        "parameters": [
          {
            "name": "geography_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Geography Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeographyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "geographies"
        ],
        "summary": "Delete Geography",
        "operationId": "delete_geography_api_v1_geographies__geography_id__delete",
        "parameters": [
          {
            "name": "geography_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Geography Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmotionCreate": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "key",
          "label"
        ],
        "title": "EmotionCreate"
      },
      "EmotionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "key",
          "label"
        ],
        "title": "EmotionRead"
      },
      "EmotionUpdate": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "key",
          "label"
        ],
        "title": "EmotionUpdate"
      },
      "GeographyCreate": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "GeographyCreate"
      },
      "GeographyRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "label",
          "description"
        ],
        "title": "GeographyRead"
      },
      "GeographyUpdate": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Key"
          },
          "label": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label"
        ],
        "title": "GeographyUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}